requirements.txt
MySQL-python==1.2.5
PyYAML==3.10
requests
beautifulsoup4